update: implements notification through SSE (Server Side Events)

+++ src/main/java/vn/com/routex/hub/notify/processor/infrastructure/kafka/event/RouteSellableEvent.java
@@ -11,7 +11,7 @@ public record RouteSellableEvent(
         String assignedBy,
         OffsetDateTime assignedAt,
         String routeStatus,
-        Integer seatCount,
+        Long seatCount,
         String creator,
         Boolean hasFloor
 ) {
